Add files via upload
Milestone 7
Compute 95% confidence intervals for the means

Interpretation:
The 95% confidence interval for the mean Sepal Length is approximately [5.786, 5.900]. This means we can be reasonably confident that the true population mean of Sepal Length falls within this interval. The confidence interval is relatively narrow, suggesting that our estimate of the mean is precise and that we have a good confidence level in the estimate.

Similarly, for Sepal Width, the 95% confidence interval is roughly [2.963, 3.151]. This interval is also relatively narrow, indicating that our estimate of the mean Sepal Width is precise, and we can be reasonably confident that the true population mean is within this range.

In both cases, the confidence intervals are relatively small, which indicates that our estimates of the means are reliable and we can trust them with a high level of confidence.